+++ client/src/components/documentCategories/DocumentCategoryList.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import axios from 'axios';
-import useProcessingAction from '../hooks/useProcessingAction';
+import useProcessingAction from '../../hooks/useProcessingAction';
 import Table from '@mui/material/Table';
 import TableBody from '@mui/material/TableBody';
 import TableCell from '@mui/material/TableCell';
@@ -15,7 +15,7 @@ import DialogContent from '@mui/material/DialogContent';
 import DialogActions from '@mui/material/DialogActions';
 import TextField from '@mui/material/TextField';
 import Card from '@mui/material/Card';
-import Typography from "@mui/material/Typography";
+import Typography from '@mui/material/Typography';
 import CardContent from '@mui/material/CardContent';
 import Grid from '@mui/material/Grid';
 import IconButton from '@mui/material/IconButton';
@@ -34,14 +34,14 @@ import RestartAltIcon from '@mui/icons-material/RestartAlt';
 import { jsPDF } from 'jspdf';
 
 function csvExport(data) {
-  const header = 'Nombre;Color fondo;Color texto';
-  const rows = data.map(t => `${t.name};${t.bgColor};${t.textColor}`);
+  const header = 'Nombre';
+  const rows = data.map(c => `${c.name}`);
   const csvContent = [header, ...rows].join('\n');
   const blob = new Blob([csvContent], { type: 'text/csv;charset=utf-8;' });
   const url = URL.createObjectURL(blob);
   const link = document.createElement('a');
   link.setAttribute('href', url);
-  link.setAttribute('download', 'tags.csv');
+  link.setAttribute('download', 'categorias_documentos.csv');
   document.body.appendChild(link);
   link.click();
   document.body.removeChild(link);
@@ -49,46 +49,50 @@ function csvExport(data) {
 
 function pdfExport(data) {
   const doc = new jsPDF();
-  doc.text('Etiquetas', 10, 10);
+  doc.text('Categorías de documentos', 10, 10);
   let y = 20;
-  data.forEach(t => {
-    doc.text(`${t.name} - ${t.bgColor} - ${t.textColor}`, 10, y);
+  data.forEach(c => {
+    doc.text(c.name, 10, y);
     y += 10;
   });
-  doc.save('tags.pdf');
+  doc.save('categorias_documentos.pdf');
 }
 
-export default function TagList({ modelId, open, onClose }) {
-  const [tags, setTags] = React.useState([]);
+export default function DocumentCategoryList({ modelId, open, onClose }) {
+  const [cats, setCats] = React.useState([]);
   const [dialogOpen, setDialogOpen] = React.useState(false);
   const [editing, setEditing] = React.useState(null);
   const [view, setView] = React.useState('table');
-  const [form, setForm] = React.useState({ name: '', bgColor: '#ffffff', textColor: '#000000' });
+  const [form, setForm] = React.useState({ name: '' });
   const [showFilters, setShowFilters] = React.useState(false);
   const [filter, setFilter] = React.useState('');
   const [sort, setSort] = React.useState({ key: 'name', dir: 'asc' });
 
   const load = async () => {
-    const res = await axios.get(`/api/models/${modelId}/tags`);
-    setTags(res.data);
+    const res = await axios.get(`/api/models/${modelId}/categoria-documentos`);
+    setCats(res.data);
   };
 
   React.useEffect(() => { if (open) load(); }, [open]);
 
   const [save, saving] = useProcessingAction(async () => {
+    if (!form.name.trim()) {
+      alert('El nombre es obligatorio');
+      return;
+    }
     if (editing) {
-      await axios.put(`/api/tags/${editing.id}`, form);
+      await axios.put(`/api/categoria-documentos/${editing.id}`, form);
     } else {
-      await axios.post(`/api/models/${modelId}/tags`, form);
+      await axios.post(`/api/models/${modelId}/categoria-documentos`, form);
     }
     setDialogOpen(false);
-    setForm({ name: '', bgColor: '#ffffff', textColor: '#000000' });
+    setForm({ name: '' });
     setEditing(null);
     load();
   });
 
   const [remove, removing] = useProcessingAction(async (id) => {
-    await axios.delete(`/api/tags/${id}`);
+    await axios.delete(`/api/categoria-documentos/${id}`);
     load();
   });
 
@@ -98,23 +102,23 @@ export default function TagList({ modelId, open, onClose }) {
     }
   };
 
-  const openEdit = (tag) => {
-    setEditing(tag);
-    setForm({ name: tag.name, bgColor: tag.bgColor, textColor: tag.textColor });
+  const openEdit = (cat) => {
+    setEditing(cat);
+    setForm({ name: cat.name });
     setDialogOpen(true);
   };
 
   const openCreate = () => {
     setEditing(null);
-    setForm({ name: '', bgColor: '#ffffff', textColor: '#000000' });
+    setForm({ name: '' });
     setDialogOpen(true);
   };
 
-  const filtered = tags.filter(t =>
-    t.name.toLowerCase().includes(filter.toLowerCase())
+  const filtered = cats.filter(c =>
+    c.name.toLowerCase().includes(filter.toLowerCase())
   );
 
-  const sorted = filtered.sort((a,b) => {
+  const sorted = filtered.sort((a,b)=>{
     const valA = a[sort.key];
     const valB = b[sort.key];
     if (valA < valB) return sort.dir === 'asc' ? -1 : 1;
@@ -128,7 +132,7 @@ export default function TagList({ modelId, open, onClose }) {
 
   return (
     <Dialog open={open} onClose={onClose} fullWidth maxWidth="md">
-      <DialogTitle>Etiquetas</DialogTitle>
+      <DialogTitle>Categorías de documentos</DialogTitle>
       <DialogContent>
         <Tooltip title={view === 'table' ? 'Vista tarjetas' : 'Vista tabla'}>
           <IconButton onClick={() => setView(view === 'table' ? 'cards' : 'table')}>
@@ -141,12 +145,12 @@ export default function TagList({ modelId, open, onClose }) {
           </IconButton>
         </Tooltip>
         <Tooltip title="Exportar CSV">
-          <IconButton onClick={() => csvExport(tags)}>
+          <IconButton onClick={() => csvExport(cats)}>
             <FileDownloadIcon />
           </IconButton>
         </Tooltip>
         <Tooltip title="Exportar PDF">
-          <IconButton onClick={() => pdfExport(tags)}>
+          <IconButton onClick={() => pdfExport(cats)}>
             <PictureAsPdfIcon />
           </IconButton>
         </Tooltip>
@@ -155,6 +159,7 @@ export default function TagList({ modelId, open, onClose }) {
             <FilterListIcon />
           </IconButton>
         </Tooltip>
+
         {showFilters && (
           <div style={{ margin: '1rem 0' }}>
           <TextField label="Buscar" value={filter} onChange={e => setFilter(e.target.value)} />
@@ -169,37 +174,25 @@ export default function TagList({ modelId, open, onClose }) {
           <TableContainer component={Paper} sx={{ mt: 2 }}>
             <Table>
           <TableHead sx={{ backgroundColor: '#f5f5f5' }}>
-                <TableRow>
+            <TableRow>
                   <TableCell onClick={() => toggleSort('name')} style={{ fontWeight: 'bold', cursor: 'pointer' }}>
                     Nombre {sort.key === 'name' && (sort.dir === 'asc' ? <ArrowUpwardIcon fontSize="small" /> : <ArrowDownwardIcon fontSize="small" />)}
-                  </TableCell>
-                  <TableCell onClick={() => toggleSort('bgColor')} style={{ fontWeight: 'bold', cursor: 'pointer' }}>
-                    Color fondo {sort.key === 'bgColor' && (sort.dir === 'asc' ? <ArrowUpwardIcon fontSize="small" /> : <ArrowDownwardIcon fontSize="small" />)}
-                  </TableCell>
-                  <TableCell onClick={() => toggleSort('textColor')} style={{ fontWeight: 'bold', cursor: 'pointer' }}>
-                    Color texto {sort.key === 'textColor' && (sort.dir === 'asc' ? <ArrowUpwardIcon fontSize="small" /> : <ArrowDownwardIcon fontSize="small" />)}
                   </TableCell>
                   <TableCell style={{ fontWeight: 'bold' }}>Acciones</TableCell>
                 </TableRow>
               </TableHead>
               <TableBody>
-                {sorted.map((tag) => (
-                  <TableRow key={tag.id}>
-                    <TableCell>{tag.name}</TableCell>
-                    <TableCell>
-                      <span style={{ backgroundColor: tag.bgColor, padding: '0.2rem 0.5rem', color: tag.textColor }}>
-                        {tag.bgColor}
-                      </span>
-                    </TableCell>
-                    <TableCell>{tag.textColor}</TableCell>
+                {sorted.map((cat) => (
+                  <TableRow key={cat.id}>
+                    <TableCell>{cat.name}</TableCell>
                     <TableCell>
                       <Tooltip title="Editar">
-                        <IconButton onClick={() => openEdit(tag)}>
+                        <IconButton onClick={() => openEdit(cat)}>
                           <EditIcon />
                         </IconButton>
                       </Tooltip>
                       <Tooltip title="Eliminar">
-                        <IconButton color="error" onClick={() => handleDelete(tag.id)} disabled={removing}>
+                        <IconButton color="error" onClick={() => handleDelete(cat.id)} disabled={removing}>
                           <DeleteIcon />
                         </IconButton>
                       </Tooltip>
@@ -211,21 +204,18 @@ export default function TagList({ modelId, open, onClose }) {
           </TableContainer>
         ) : (
           <Grid container spacing={2} sx={{ mt: 2 }}>
-            {sorted.map((tag) => (
-              <Grid item xs={12} md={4} key={tag.id}>
+            {sorted.map((cat) => (
+              <Grid item xs={12} md={4} key={cat.id}>
                 <Card>
                   <CardContent>
-                    <Typography variant="h6">{tag.name}</Typography>
-                    <div style={{ backgroundColor: tag.bgColor, color: tag.textColor, padding: '0.5rem' }}>
-                      {tag.name}
-                    </div>
+                    <Typography variant="h6">{cat.name}</Typography>
                     <Tooltip title="Editar">
-                      <IconButton onClick={() => openEdit(tag)}>
+                      <IconButton onClick={() => openEdit(cat)}>
                         <EditIcon />
                       </IconButton>
                     </Tooltip>
                     <Tooltip title="Eliminar">
-                      <IconButton color="error" onClick={() => handleDelete(tag.id)} disabled={removing}>
+                      <IconButton color="error" onClick={() => handleDelete(cat.id)} disabled={removing}>
                         <DeleteIcon />
                       </IconButton>
                     </Tooltip>
@@ -236,11 +226,9 @@ export default function TagList({ modelId, open, onClose }) {
           </Grid>
         )}
         <Dialog open={dialogOpen} onClose={() => setDialogOpen(false)}>
-          <DialogTitle>{editing ? 'Editar' : 'Nueva'} etiqueta</DialogTitle>
+          <DialogTitle>{editing ? 'Editar' : 'Nueva'} categoría de documentos</DialogTitle>
           <DialogContent>
-            <TextField required label="Nombre" value={form.name} onChange={(e) => setForm({ ...form, name: e.target.value })} fullWidth />
-            <TextField required label="Color fondo" type="color" value={form.bgColor} onChange={(e) => setForm({ ...form, bgColor: e.target.value })} fullWidth sx={{ mt: 2 }} />
-            <TextField required label="Color texto" type="color" value={form.textColor} onChange={(e) => setForm({ ...form, textColor: e.target.value })} fullWidth sx={{ mt: 2 }} />
+            <TextField required label="Nombre" value={form.name} onChange={(e) => setForm({ name: e.target.value })} fullWidth />
           </DialogContent>
           <DialogActions>
             <Button onClick={() => setDialogOpen(false)} disabled={saving}>Cancelar</Button>
