+++ client/src/components/parameters/ParameterList.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import axios from 'axios';
-import useProcessingAction from '../hooks/useProcessingAction';
+import useProcessingAction from '../../hooks/useProcessingAction';
 import Table from '@mui/material/Table';
 import TableBody from '@mui/material/TableBody';
 import TableCell from '@mui/material/TableCell';
