+++ client/src/components/roles/RoleList.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import axios from 'axios';
-import useProcessingAction from '../hooks/useProcessingAction';
+import useProcessingAction from '../../hooks/useProcessingAction';
 import Table from '@mui/material/Table';
 import TableBody from '@mui/material/TableBody';
 import TableCell from '@mui/material/TableCell';
@@ -15,7 +15,7 @@ import DialogContent from '@mui/material/DialogContent';
 import DialogActions from '@mui/material/DialogActions';
 import TextField from '@mui/material/TextField';
 import Card from '@mui/material/Card';
-import Typography from '@mui/material/Typography';
+import Typography from "@mui/material/Typography";
 import CardContent from '@mui/material/CardContent';
 import Grid from '@mui/material/Grid';
 import IconButton from '@mui/material/IconButton';
@@ -34,14 +34,14 @@ import RestartAltIcon from '@mui/icons-material/RestartAlt';
 import { jsPDF } from 'jspdf';
 
 function csvExport(data) {
-  const header = 'Nombre';
-  const rows = data.map(c => `${c.name}`);
+  const header = 'Nombre;Orden';
+  const rows = data.map(r => `${r.name};${r.order}`);
   const csvContent = [header, ...rows].join('\n');
   const blob = new Blob([csvContent], { type: 'text/csv;charset=utf-8;' });
   const url = URL.createObjectURL(blob);
   const link = document.createElement('a');
   link.setAttribute('href', url);
-  link.setAttribute('download', 'categorias_documentos.csv');
+  link.setAttribute('download', 'roles.csv');
   document.body.appendChild(link);
   link.click();
   document.body.removeChild(link);
@@ -49,76 +49,72 @@ function csvExport(data) {
 
 function pdfExport(data) {
   const doc = new jsPDF();
-  doc.text('Categorías de documentos', 10, 10);
+  doc.text('Roles', 10, 10);
   let y = 20;
-  data.forEach(c => {
-    doc.text(c.name, 10, y);
+  data.forEach(r => {
+    doc.text(`${r.order} - ${r.name}`, 10, y);
     y += 10;
   });
-  doc.save('categorias_documentos.pdf');
+  doc.save('roles.pdf');
 }
 
-export default function DocumentCategoryList({ modelId, open, onClose }) {
-  const [cats, setCats] = React.useState([]);
+export default function RoleList({ teamId, teamName, open, onClose }) {
+  const [roles, setRoles] = React.useState([]);
   const [dialogOpen, setDialogOpen] = React.useState(false);
   const [editing, setEditing] = React.useState(null);
   const [view, setView] = React.useState('table');
-  const [form, setForm] = React.useState({ name: '' });
+  const [form, setForm] = React.useState({ name: '', order: 0 });
   const [showFilters, setShowFilters] = React.useState(false);
   const [filter, setFilter] = React.useState('');
-  const [sort, setSort] = React.useState({ key: 'name', dir: 'asc' });
+  const [sort, setSort] = React.useState({ key: 'order', dir: 'asc' });
 
   const load = async () => {
-    const res = await axios.get(`/api/models/${modelId}/categoria-documentos`);
-    setCats(res.data);
+    const res = await axios.get(`/api/teams/${teamId}/roles`);
+    setRoles(res.data);
   };
 
   React.useEffect(() => { if (open) load(); }, [open]);
 
-  const [save, saving] = useProcessingAction(async () => {
-    if (!form.name.trim()) {
-      alert('El nombre es obligatorio');
-      return;
-    }
+  const [saveRole, saving] = useProcessingAction(async () => {
     if (editing) {
-      await axios.put(`/api/categoria-documentos/${editing.id}`, form);
+      await axios.put(`/api/roles/${editing.id}`, form);
     } else {
-      await axios.post(`/api/models/${modelId}/categoria-documentos`, form);
+      await axios.post(`/api/teams/${teamId}/roles`, form);
     }
     setDialogOpen(false);
-    setForm({ name: '' });
+    setForm({ name: '', order: 0 });
     setEditing(null);
     load();
   });
 
-  const [remove, removing] = useProcessingAction(async (id) => {
-    await axios.delete(`/api/categoria-documentos/${id}`);
+  const [removeRole, removing] = useProcessingAction(async (id) => {
+    await axios.delete(`/api/roles/${id}`);
     load();
   });
 
   const handleDelete = (id) => {
     if (window.confirm('¿Eliminar elemento?')) {
-      remove(id);
+      removeRole(id);
     }
   };
 
-  const openEdit = (cat) => {
-    setEditing(cat);
-    setForm({ name: cat.name });
+  const openEdit = (role) => {
+    setEditing(role);
+    setForm({ name: role.name, order: role.order });
     setDialogOpen(true);
   };
 
   const openCreate = () => {
     setEditing(null);
-    setForm({ name: '' });
+    setForm({ name: '', order: 0 });
     setDialogOpen(true);
   };
 
-  const filtered = cats.filter(c =>
-    c.name.toLowerCase().includes(filter.toLowerCase())
+  const filtered = roles.filter(r =>
+    r.name.toLowerCase().includes(filter.toLowerCase())
   );
 
-  const sorted = filtered.sort((a,b)=>{
+  const sorted = filtered.sort((a,b) => {
     const valA = a[sort.key];
     const valB = b[sort.key];
     if (valA < valB) return sort.dir === 'asc' ? -1 : 1;
@@ -132,25 +128,25 @@ export default function DocumentCategoryList({ modelId, open, onClose }) {
 
   return (
     <Dialog open={open} onClose={onClose} fullWidth maxWidth="md">
-      <DialogTitle>Categorías de documentos</DialogTitle>
+      <DialogTitle>{`Roles del equipo ${teamName}`}</DialogTitle>
       <DialogContent>
         <Tooltip title={view === 'table' ? 'Vista tarjetas' : 'Vista tabla'}>
           <IconButton onClick={() => setView(view === 'table' ? 'cards' : 'table')}>
             {view === 'table' ? <ViewModuleIcon /> : <TableRowsIcon />}
           </IconButton>
         </Tooltip>
-        <Tooltip title="Nueva">
+        <Tooltip title="Nuevo">
           <IconButton onClick={openCreate}>
             <AddIcon />
           </IconButton>
         </Tooltip>
         <Tooltip title="Exportar CSV">
-          <IconButton onClick={() => csvExport(cats)}>
+          <IconButton onClick={() => csvExport(roles)}>
             <FileDownloadIcon />
           </IconButton>
         </Tooltip>
         <Tooltip title="Exportar PDF">
-          <IconButton onClick={() => pdfExport(cats)}>
+          <IconButton onClick={() => pdfExport(roles)}>
             <PictureAsPdfIcon />
           </IconButton>
         </Tooltip>
@@ -159,7 +155,6 @@ export default function DocumentCategoryList({ modelId, open, onClose }) {
             <FilterListIcon />
           </IconButton>
         </Tooltip>
-
         {showFilters && (
           <div style={{ margin: '1rem 0' }}>
           <TextField label="Buscar" value={filter} onChange={e => setFilter(e.target.value)} />
@@ -174,7 +169,10 @@ export default function DocumentCategoryList({ modelId, open, onClose }) {
           <TableContainer component={Paper} sx={{ mt: 2 }}>
             <Table>
           <TableHead sx={{ backgroundColor: '#f5f5f5' }}>
-            <TableRow>
+                <TableRow>
+                  <TableCell onClick={() => toggleSort('order')} style={{ fontWeight: 'bold', cursor: 'pointer' }}>
+                    Orden {sort.key === 'order' && (sort.dir === 'asc' ? <ArrowUpwardIcon fontSize="small" /> : <ArrowDownwardIcon fontSize="small" />)}
+                  </TableCell>
                   <TableCell onClick={() => toggleSort('name')} style={{ fontWeight: 'bold', cursor: 'pointer' }}>
                     Nombre {sort.key === 'name' && (sort.dir === 'asc' ? <ArrowUpwardIcon fontSize="small" /> : <ArrowDownwardIcon fontSize="small" />)}
                   </TableCell>
@@ -182,17 +180,18 @@ export default function DocumentCategoryList({ modelId, open, onClose }) {
                 </TableRow>
               </TableHead>
               <TableBody>
-                {sorted.map((cat) => (
-                  <TableRow key={cat.id}>
-                    <TableCell>{cat.name}</TableCell>
+                {sorted.map((role) => (
+                  <TableRow key={role.id}>
+                    <TableCell>{role.order}</TableCell>
+                    <TableCell>{role.name}</TableCell>
                     <TableCell>
                       <Tooltip title="Editar">
-                        <IconButton onClick={() => openEdit(cat)}>
+                        <IconButton onClick={() => openEdit(role)}>
                           <EditIcon />
                         </IconButton>
                       </Tooltip>
                       <Tooltip title="Eliminar">
-                        <IconButton color="error" onClick={() => handleDelete(cat.id)} disabled={removing}>
+                        <IconButton color="error" onClick={() => handleDelete(role.id)} disabled={removing}>
                           <DeleteIcon />
                         </IconButton>
                       </Tooltip>
@@ -204,18 +203,18 @@ export default function DocumentCategoryList({ modelId, open, onClose }) {
           </TableContainer>
         ) : (
           <Grid container spacing={2} sx={{ mt: 2 }}>
-            {sorted.map((cat) => (
-              <Grid item xs={12} md={4} key={cat.id}>
+            {sorted.map((role) => (
+              <Grid item xs={12} md={4} key={role.id}>
                 <Card>
                   <CardContent>
-                    <Typography variant="h6">{cat.name}</Typography>
+                    <Typography variant="h6">{role.order} - {role.name}</Typography>
                     <Tooltip title="Editar">
-                      <IconButton onClick={() => openEdit(cat)}>
+                      <IconButton onClick={() => openEdit(role)}>
                         <EditIcon />
                       </IconButton>
                     </Tooltip>
                     <Tooltip title="Eliminar">
-                      <IconButton color="error" onClick={() => handleDelete(cat.id)} disabled={removing}>
+                      <IconButton color="error" onClick={() => handleDelete(role.id)} disabled={removing}>
                         <DeleteIcon />
                       </IconButton>
                     </Tooltip>
@@ -226,13 +225,14 @@ export default function DocumentCategoryList({ modelId, open, onClose }) {
           </Grid>
         )}
         <Dialog open={dialogOpen} onClose={() => setDialogOpen(false)}>
-          <DialogTitle>{editing ? 'Editar' : 'Nueva'} categoría de documentos</DialogTitle>
+          <DialogTitle>{editing ? 'Editar' : 'Nuevo'} rol</DialogTitle>
           <DialogContent>
-            <TextField required label="Nombre" value={form.name} onChange={(e) => setForm({ name: e.target.value })} fullWidth />
+            <TextField required label="Nombre" value={form.name} onChange={(e) => setForm({ ...form, name: e.target.value })} fullWidth />
+            <TextField required label="Orden" type="number" value={form.order} onChange={(e) => setForm({ ...form, order: parseInt(e.target.value, 10) })} fullWidth sx={{ mt: 2 }} />
           </DialogContent>
           <DialogActions>
             <Button onClick={() => setDialogOpen(false)} disabled={saving}>Cancelar</Button>
-            <Button onClick={save} disabled={saving}>Guardar</Button>
+            <Button onClick={saveRole} disabled={saving}>Guardar</Button>
           </DialogActions>
         </Dialog>
       </DialogContent>
