+++ client/src/components/tags/TagList.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import axios from 'axios';
-import useProcessingAction from '../hooks/useProcessingAction';
+import useProcessingAction from '../../hooks/useProcessingAction';
 import Table from '@mui/material/Table';
 import TableBody from '@mui/material/TableBody';
 import TableCell from '@mui/material/TableCell';
@@ -34,14 +34,14 @@ import RestartAltIcon from '@mui/icons-material/RestartAlt';
 import { jsPDF } from 'jspdf';
 
 function csvExport(data) {
-  const header = 'Nombre;Orden';
-  const rows = data.map(r => `${r.name};${r.order}`);
+  const header = 'Nombre;Color fondo;Color texto';
+  const rows = data.map(t => `${t.name};${t.bgColor};${t.textColor}`);
   const csvContent = [header, ...rows].join('\n');
   const blob = new Blob([csvContent], { type: 'text/csv;charset=utf-8;' });
   const url = URL.createObjectURL(blob);
   const link = document.createElement('a');
   link.setAttribute('href', url);
-  link.setAttribute('download', 'roles.csv');
+  link.setAttribute('download', 'tags.csv');
   document.body.appendChild(link);
   link.click();
   document.body.removeChild(link);
@@ -49,69 +49,69 @@ function csvExport(data) {
 
 function pdfExport(data) {
   const doc = new jsPDF();
-  doc.text('Roles', 10, 10);
+  doc.text('Etiquetas', 10, 10);
   let y = 20;
-  data.forEach(r => {
-    doc.text(`${r.order} - ${r.name}`, 10, y);
+  data.forEach(t => {
+    doc.text(`${t.name} - ${t.bgColor} - ${t.textColor}`, 10, y);
     y += 10;
   });
-  doc.save('roles.pdf');
+  doc.save('tags.pdf');
 }
 
-export default function RoleList({ teamId, teamName, open, onClose }) {
-  const [roles, setRoles] = React.useState([]);
+export default function TagList({ modelId, open, onClose }) {
+  const [tags, setTags] = React.useState([]);
   const [dialogOpen, setDialogOpen] = React.useState(false);
   const [editing, setEditing] = React.useState(null);
   const [view, setView] = React.useState('table');
-  const [form, setForm] = React.useState({ name: '', order: 0 });
+  const [form, setForm] = React.useState({ name: '', bgColor: '#ffffff', textColor: '#000000' });
   const [showFilters, setShowFilters] = React.useState(false);
   const [filter, setFilter] = React.useState('');
-  const [sort, setSort] = React.useState({ key: 'order', dir: 'asc' });
+  const [sort, setSort] = React.useState({ key: 'name', dir: 'asc' });
 
   const load = async () => {
-    const res = await axios.get(`/api/teams/${teamId}/roles`);
-    setRoles(res.data);
+    const res = await axios.get(`/api/models/${modelId}/tags`);
+    setTags(res.data);
   };
 
   React.useEffect(() => { if (open) load(); }, [open]);
 
-  const [saveRole, saving] = useProcessingAction(async () => {
+  const [save, saving] = useProcessingAction(async () => {
     if (editing) {
-      await axios.put(`/api/roles/${editing.id}`, form);
+      await axios.put(`/api/tags/${editing.id}`, form);
     } else {
-      await axios.post(`/api/teams/${teamId}/roles`, form);
+      await axios.post(`/api/models/${modelId}/tags`, form);
     }
     setDialogOpen(false);
-    setForm({ name: '', order: 0 });
+    setForm({ name: '', bgColor: '#ffffff', textColor: '#000000' });
     setEditing(null);
     load();
   });
 
-  const [removeRole, removing] = useProcessingAction(async (id) => {
-    await axios.delete(`/api/roles/${id}`);
+  const [remove, removing] = useProcessingAction(async (id) => {
+    await axios.delete(`/api/tags/${id}`);
     load();
   });
 
   const handleDelete = (id) => {
     if (window.confirm('¿Eliminar elemento?')) {
-      removeRole(id);
+      remove(id);
     }
   };
 
-  const openEdit = (role) => {
-    setEditing(role);
-    setForm({ name: role.name, order: role.order });
+  const openEdit = (tag) => {
+    setEditing(tag);
+    setForm({ name: tag.name, bgColor: tag.bgColor, textColor: tag.textColor });
     setDialogOpen(true);
   };
 
   const openCreate = () => {
     setEditing(null);
-    setForm({ name: '', order: 0 });
+    setForm({ name: '', bgColor: '#ffffff', textColor: '#000000' });
     setDialogOpen(true);
   };
 
-  const filtered = roles.filter(r =>
-    r.name.toLowerCase().includes(filter.toLowerCase())
+  const filtered = tags.filter(t =>
+    t.name.toLowerCase().includes(filter.toLowerCase())
   );
 
   const sorted = filtered.sort((a,b) => {
@@ -128,25 +128,25 @@ export default function RoleList({ teamId, teamName, open, onClose }) {
 
   return (
     <Dialog open={open} onClose={onClose} fullWidth maxWidth="md">
-      <DialogTitle>{`Roles del equipo ${teamName}`}</DialogTitle>
+      <DialogTitle>Etiquetas</DialogTitle>
       <DialogContent>
         <Tooltip title={view === 'table' ? 'Vista tarjetas' : 'Vista tabla'}>
           <IconButton onClick={() => setView(view === 'table' ? 'cards' : 'table')}>
             {view === 'table' ? <ViewModuleIcon /> : <TableRowsIcon />}
           </IconButton>
         </Tooltip>
-        <Tooltip title="Nuevo">
+        <Tooltip title="Nueva">
           <IconButton onClick={openCreate}>
             <AddIcon />
           </IconButton>
         </Tooltip>
         <Tooltip title="Exportar CSV">
-          <IconButton onClick={() => csvExport(roles)}>
+          <IconButton onClick={() => csvExport(tags)}>
             <FileDownloadIcon />
           </IconButton>
         </Tooltip>
         <Tooltip title="Exportar PDF">
-          <IconButton onClick={() => pdfExport(roles)}>
+          <IconButton onClick={() => pdfExport(tags)}>
             <PictureAsPdfIcon />
           </IconButton>
         </Tooltip>
@@ -170,28 +170,36 @@ export default function RoleList({ teamId, teamName, open, onClose }) {
             <Table>
           <TableHead sx={{ backgroundColor: '#f5f5f5' }}>
                 <TableRow>
-                  <TableCell onClick={() => toggleSort('order')} style={{ fontWeight: 'bold', cursor: 'pointer' }}>
-                    Orden {sort.key === 'order' && (sort.dir === 'asc' ? <ArrowUpwardIcon fontSize="small" /> : <ArrowDownwardIcon fontSize="small" />)}
-                  </TableCell>
                   <TableCell onClick={() => toggleSort('name')} style={{ fontWeight: 'bold', cursor: 'pointer' }}>
                     Nombre {sort.key === 'name' && (sort.dir === 'asc' ? <ArrowUpwardIcon fontSize="small" /> : <ArrowDownwardIcon fontSize="small" />)}
+                  </TableCell>
+                  <TableCell onClick={() => toggleSort('bgColor')} style={{ fontWeight: 'bold', cursor: 'pointer' }}>
+                    Color fondo {sort.key === 'bgColor' && (sort.dir === 'asc' ? <ArrowUpwardIcon fontSize="small" /> : <ArrowDownwardIcon fontSize="small" />)}
+                  </TableCell>
+                  <TableCell onClick={() => toggleSort('textColor')} style={{ fontWeight: 'bold', cursor: 'pointer' }}>
+                    Color texto {sort.key === 'textColor' && (sort.dir === 'asc' ? <ArrowUpwardIcon fontSize="small" /> : <ArrowDownwardIcon fontSize="small" />)}
                   </TableCell>
                   <TableCell style={{ fontWeight: 'bold' }}>Acciones</TableCell>
                 </TableRow>
               </TableHead>
               <TableBody>
-                {sorted.map((role) => (
-                  <TableRow key={role.id}>
-                    <TableCell>{role.order}</TableCell>
-                    <TableCell>{role.name}</TableCell>
+                {sorted.map((tag) => (
+                  <TableRow key={tag.id}>
+                    <TableCell>{tag.name}</TableCell>
+                    <TableCell>
+                      <span style={{ backgroundColor: tag.bgColor, padding: '0.2rem 0.5rem', color: tag.textColor }}>
+                        {tag.bgColor}
+                      </span>
+                    </TableCell>
+                    <TableCell>{tag.textColor}</TableCell>
                     <TableCell>
                       <Tooltip title="Editar">
-                        <IconButton onClick={() => openEdit(role)}>
+                        <IconButton onClick={() => openEdit(tag)}>
                           <EditIcon />
                         </IconButton>
                       </Tooltip>
                       <Tooltip title="Eliminar">
-                        <IconButton color="error" onClick={() => handleDelete(role.id)} disabled={removing}>
+                        <IconButton color="error" onClick={() => handleDelete(tag.id)} disabled={removing}>
                           <DeleteIcon />
                         </IconButton>
                       </Tooltip>
@@ -203,18 +211,21 @@ export default function RoleList({ teamId, teamName, open, onClose }) {
           </TableContainer>
         ) : (
           <Grid container spacing={2} sx={{ mt: 2 }}>
-            {sorted.map((role) => (
-              <Grid item xs={12} md={4} key={role.id}>
+            {sorted.map((tag) => (
+              <Grid item xs={12} md={4} key={tag.id}>
                 <Card>
                   <CardContent>
-                    <Typography variant="h6">{role.order} - {role.name}</Typography>
+                    <Typography variant="h6">{tag.name}</Typography>
+                    <div style={{ backgroundColor: tag.bgColor, color: tag.textColor, padding: '0.5rem' }}>
+                      {tag.name}
+                    </div>
                     <Tooltip title="Editar">
-                      <IconButton onClick={() => openEdit(role)}>
+                      <IconButton onClick={() => openEdit(tag)}>
                         <EditIcon />
                       </IconButton>
                     </Tooltip>
                     <Tooltip title="Eliminar">
-                      <IconButton color="error" onClick={() => handleDelete(role.id)} disabled={removing}>
+                      <IconButton color="error" onClick={() => handleDelete(tag.id)} disabled={removing}>
                         <DeleteIcon />
                       </IconButton>
                     </Tooltip>
@@ -225,14 +236,15 @@ export default function RoleList({ teamId, teamName, open, onClose }) {
           </Grid>
         )}
         <Dialog open={dialogOpen} onClose={() => setDialogOpen(false)}>
-          <DialogTitle>{editing ? 'Editar' : 'Nuevo'} rol</DialogTitle>
+          <DialogTitle>{editing ? 'Editar' : 'Nueva'} etiqueta</DialogTitle>
           <DialogContent>
             <TextField required label="Nombre" value={form.name} onChange={(e) => setForm({ ...form, name: e.target.value })} fullWidth />
-            <TextField required label="Orden" type="number" value={form.order} onChange={(e) => setForm({ ...form, order: parseInt(e.target.value, 10) })} fullWidth sx={{ mt: 2 }} />
+            <TextField required label="Color fondo" type="color" value={form.bgColor} onChange={(e) => setForm({ ...form, bgColor: e.target.value })} fullWidth sx={{ mt: 2 }} />
+            <TextField required label="Color texto" type="color" value={form.textColor} onChange={(e) => setForm({ ...form, textColor: e.target.value })} fullWidth sx={{ mt: 2 }} />
           </DialogContent>
           <DialogActions>
             <Button onClick={() => setDialogOpen(false)} disabled={saving}>Cancelar</Button>
-            <Button onClick={saveRole} disabled={saving}>Guardar</Button>
+            <Button onClick={save} disabled={saving}>Guardar</Button>
           </DialogActions>
         </Dialog>
       </DialogContent>
