+++ client/src/components/teams/TeamList.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import axios from 'axios';
-import useProcessingAction from '../hooks/useProcessingAction';
+import useProcessingAction from '../../hooks/useProcessingAction';
 import Table from '@mui/material/Table';
 import TableBody from '@mui/material/TableBody';
 import TableCell from '@mui/material/TableCell';
@@ -34,7 +34,7 @@ import GroupIcon from '@mui/icons-material/Group';
 import Badge from '@mui/material/Badge';
 import RestartAltIcon from '@mui/icons-material/RestartAlt';
 import { jsPDF } from 'jspdf';
-import RoleList from './RoleList';
+import RoleList from '../roles/RoleList';
 
 function csvExport(data) {
   const header = 'Nombre;Orden';
