+++ server/models/Model.js
@@ -1,0 +1,16 @@
+module.exports = (sequelize, DataTypes) => {
+  return sequelize.define('Model', {
+    name: {
+      type: DataTypes.STRING,
+      allowNull: false,
+    },
+    author: {
+      type: DataTypes.STRING,
+      allowNull: false,
+    },
+    parentId: {
+      type: DataTypes.INTEGER,
+      allowNull: true,
+    }
+  });
+};
