+++ server/models/Node.js
@@ -1,0 +1,36 @@
+module.exports = (sequelize, DataTypes) => {
+  return sequelize.define('Node', {
+    name: {
+      type: DataTypes.STRING,
+      allowNull: false,
+    },
+    order: {
+      type: DataTypes.INTEGER,
+      allowNull: false,
+      defaultValue: 0,
+    },
+    codePattern: {
+      type: DataTypes.STRING(5),
+      allowNull: false,
+      defaultValue: 'ORDER',
+    },
+    code: {
+      type: DataTypes.STRING,
+      allowNull: false,
+      defaultValue: '',
+    },
+    description: {
+      type: DataTypes.TEXT,
+      allowNull: false,
+      defaultValue: '',
+    },
+    modelId: {
+      type: DataTypes.INTEGER,
+      allowNull: false,
+    },
+    parentId: {
+      type: DataTypes.INTEGER,
+      allowNull: true,
+    },
+  });
+};
