+++ server/models/NodeRasci.js
@@ -1,0 +1,8 @@
+module.exports = (sequelize, DataTypes) => {
+  return sequelize.define('NodeRasci', {
+    responsibilities: {
+      type: DataTypes.STRING,
+      allowNull: false,
+    },
+  });
+};
