+++ server/models/NodeTag.js
@@ -1,0 +1,3 @@
+module.exports = (sequelize) => {
+  return sequelize.define('NodeTag', {});
+};
