+++ server/models/Team.js
@@ -1,0 +1,16 @@
+module.exports = (sequelize, DataTypes) => {
+  return sequelize.define('Team', {
+    name: {
+      type: DataTypes.STRING,
+      allowNull: false,
+    },
+    order: {
+      type: DataTypes.INTEGER,
+      allowNull: false,
+    },
+    modelId: {
+      type: DataTypes.INTEGER,
+      allowNull: false,
+    },
+  });
+};
